Update 03-02-or.html
I do not think native English speakers know to look for "hausen" or "roda". Maybe explain why you would use those. I only think to try ```|bach``` or ```bach|``` but don't know what to put on the other side of the logic operator. 

The instructions say to look for words with "bach" in them. Why would I look for "hausen" or "roda" ?